Add shift DashboardPage to src/pages
